tests/requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colorlog==5.0.1
curlify==2.2.1
docopt==0.6.2
graphql-core==3.1.5
hypothesis==6.13.6
hypothesis-graphql==0.5.0
hypothesis-jsonschema==0.19.2
idna==2.10
iniconfig==1.1.1
jmespath==0.10.0
jsonschema==3.2.0
junit-xml==1.9
multidict==5.1.0
packaging==20.9
paho-mqtt==1.5.1
pbr==5.6.0
pluggy==0.13.1
py==1.10.0
PyJWT==1.7.1
pykwalify==1.7.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.4
pytest-subtests==0.4.0
python-box==5.3.0
python-dateutil==2.8.1
PyYAML==5.4.1
requests==2.25.1
schemathesis==3.7.1
six==1.16.0
sortedcontainers==2.4.0
starlette==0.14.2
stevedore==3.3.0
tavern==1.14.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
Werkzeug==2.0.1
yarl==1.6.3
